pyproject.toml
[project]
name = "valrdotcom-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "asgiref (>=3.8.1,<4.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "cffi (>=1.17.1,<2.0.0)",
    "charset-normalizer (>=3.4.1,<4.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "defusedxml (>=0.7.1,<0.8.0)",
    "django (>=5.1.7,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-templated-mail (>=1.1.1,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djoser (>=2.3.1,<3.0.0)",
    "oauthlib (>=3.2.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python3-openid (>=3.2.0,<4.0.0)",
    "pycparser (>=2.22,<3.0)",
    "pytz (>=2025.2,<2026.0)",
    "idna (>=3.10,<4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
    "social-auth-app-django (>=5.4.3,<6.0.0)",
    "social-auth-core (>=4.5.6,<5.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "stripe (>=11.6.0,<12.0.0)",
    "typing-extensions (>=4.13.0,<5.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "djangorestframework-simplejwt (==5.4.0)",
    "pyjwt (>=1.7.1,<2.10.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
